Image recognize v1
Basic activation function. Image decording. 210 45*62 images of nums and 10 5*9 images of nums.

+++ heron.h
@@ -15,21 +15,21 @@ public:
 
     //int mLinksOut[];// this may delete
     //int mLinksIn[];// this may delete
-//    float mNext, mRealy; // this may delete
-//    float l; /// summ for enters l
-//    float o;///  heron output
-    float mLayerError = 0;
+//    double mNext, mRealy; // this may delete
+//    double l; /// summ for enters l
+//    double o;///  heron output
+    double mLayerError = 0;
 
     QString toQString();
     Link getLinkForHeron(int layer, int heron);
-    void addLink(Heron *heron, float weight);
+    void addLink(Heron *heron, double weight);
     int mIdLayer;
     int mId;
-    QList<float> mInputs;
-    float mOutput = 0;
+    QList<double> mInputs;
+    double mOutput = 0;
 
 // i coment strings there we dont use now!
-//    float lastDWeight; // delta between weights
+//    double lastDWeight; // delta between weights
     QList<Link> mLinks = QList<Link>();// output links (there sender this)
 };
 class Link
@@ -38,8 +38,8 @@ public:
 
     Heron* mSender;
     Heron* mRecever;
-    float mWeight = 1;
-    float mLastDWeight = 0;
+    double mWeight = 1;
+    double mLastDWeight = 0;
     QString toQString(){
         return QString("Link for Heron {"+QString::number(mSender->mIdLayer)
                        +QString::number(mSender->mId)+"} to {"
@@ -47,6 +47,6 @@ public:
                        +QString::number(mRecever->mId)+ "} weight = "
                        +QString::number(mWeight));
     }
-    Link(Heron* sender,Heron* recever,float weight);
+    Link(Heron* sender,Heron* recever,double weight);
 };
 #endif // HERON_H
